Refactor Action
* Renames ActionDescriptor to Action.
* Renames the existing Action class to EventListener.
* Replaces InlineActionObserver with ActionObserver, built on TokenObserver.
