+++ packages/@stimulus/core/index.ts
@@ -1,5 +1,3 @@
-export { Action } from "./src/action"
-export { ActionDescriptor } from "./src/action_descriptor"
 export { Application } from "./src/application"
 export { Context } from "./src/context"
 export { Controller, ControllerConstructor } from "./src/controller"
